modules/neuranlp_agent/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
openai-whisper
chromadb
tiktoken
pyttsx3
pydub
requests
python-dotenv
ollama